HC 20231029 refactoring
