+++ blogfront/src/components/EditTag/EditTag.js
@@ -1,12 +1,9 @@
-
 import InputBox from './InputBox';
 import SubBtn from './SubBtn';
 import styled from 'styled-components'
-
-
 import img from './Source/bgimg.png'
 import MappingContainer from '../Setting/MappingContainer';
-import BackToSet from '../Setting/BackToSet';
+import HeaderBox from '../Setting/HeaderBox';
 
 let Container = styled.div`
 
@@ -14,7 +11,7 @@ let Container = styled.div`
     justify-content: center;
     align-items: center;
     height: 100vh;
-    width: 100vw;
+    width: calc(100vw - 7rem);
     background-image: url(${img});
     background-size: cover; 
 
@@ -32,43 +29,7 @@ let SettingBox = styled.div`
 
 `
 
-let ImgBox = styled.div`
-
-flex-grow: 3;
-display: flex;
-align-items: center;
-justify-content: center;
-
-`
-
-let H = styled.div`
-
-flex-grow: 7;
-display: flex;
-align-items: center;
-justify-content: center;
-font-size: 20px;
-
-
-`
-let HeaderContainer = styled.div`
-
-height: 10vh;
-justify-content: center;
-width: 100%;
-display: flex;
-flex-direction: row;
-
-
-`
-
-
-
-
 function EditTag() {
-
-
-
 
     return (
 
@@ -76,26 +37,18 @@ function EditTag() {
             <Container>
                 {/* setting할 수 있는 메뉴를 불러오는데 props를 이용하여 설정 */}
 
-
                 <SettingBox>
+
                     <MappingContainer />
-                    <HeaderContainer>
-                        <H>선호 태그 재설정</H>
-                        <ImgBox>
-                            <BackToSet />
-                        </ImgBox>
-                    </HeaderContainer>
+                    <HeaderBox headerText="선호태그 변경"/>
                     <InputBox />
-
                     <SubBtn />
-
                 </SettingBox>
 
             </Container>
         </>
+
     )
-
-
 
 }
 export default EditTag;
